+++ sdmx/model/common.py
@@ -99,14 +99,33 @@ class AnnotableArtefact(BaseModel):
 
         raise KeyError(attrib)
 
+    def eval_annotation(self, id: str, globals=None):
+        """Retrieve the annotation with the given `id` and :func:`eval` its contents.
+
+        This can be used for unpacking Python values (e.g. :class:`dict`) stored as an
+        annotation on an AnnotableArtefact (e.g. :class:`~sdmx.model.Code`).
+
+        Returns :obj:`None` if no attribute exists with the given `id`.
+        """
+        try:
+            value = str(self.get_annotation(id=id).text)
+        except KeyError:  # No such attribute
+            return None
+
+        try:
+            return eval(value, globals or {})
+        except Exception as e:  # Something that can't be eval()'d, e.g. a plain string
+            log.debug(f"Could not eval({value!r}): {e}")
+            return value
+
 
 class IdentifiableArtefact(AnnotableArtefact):
     #: Unique identifier of the object.
     id: str = MissingID
     #: Universal resource identifier that may or may not be resolvable.
     uri: Optional[str] = None
-    #: Universal resource name. For use in SDMX registries; all registered
-    #: objects have a URN.
+    #: Universal resource name. For use in SDMX registries; all registered objects have
+    #: a URN.
     urn: Optional[str] = None
 
     urn_group: Dict = dict()
@@ -275,8 +294,7 @@ class MaintainableArtefact(VersionableArtefact):
     #: :obj:`True` if the content of the object is held externally; i.e., not
     #: the current :class:`Message`.
     is_external_reference: Optional[bool] = None
-    #: URL of an SDMX-compliant web service from which the object can be
-    #: retrieved.
+    #: URL of an SDMX-compliant web service from which the object can be retrieved.
     service_url: Optional[str] = None
     #: URL of an SDMX-ML document containing the object.
     structure_url: Optional[str] = None